yzy_web/requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
dynaconf==2.2.3
idna==2.9
importlib-metadata==1.5.0
jsonschema==3.2.0
psutil==5.6.7
PyMySQL==0.9.3
pyrsistent==0.15.7
pytz==2019.3
python-box=3.4.6
python-dotenv=0.13.0
requests==2.23.0
six==1.14.0
toml=0.10.1
urllib3==1.25.8
zipp==3.1.0

